move scenario structure to separate file
